Move tasks into seperate command classes instead of one robofile
